frontend/requirements.txt
flask
flask-cors
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
